honksquad: refactor: relocate drop-to-release-pull out of PullingSystem
The fork's drop-to-release-pull behavior (drop the virtual item in your
hand to stop pulling) lived as three HONK blocks inside upstream
PullingSystem.cs. It doesn't need to. TryStopPull is already public and
the handler has no private dependencies, so the whole subscription moves
to a fork-owned VirtualItemPullReleaseSystem under RussStation/Pulling.

Shrinks the fork's footprint in the most rebase-volatile file in the
codebase. Behavior-preserving. Part of issue #441 P0.1.

The force: true argument on TryStopPull stays until P0.4 lands and
replaces the force parameter with a dedicated PullReleaseRequestedEvent.
